Move rust runtime api tests to the additional internal crate
