+++ runtimes/rust/Cargo.toml
@@ -30,8 +30,6 @@ serde = "1.0.101"
 serde_derive = "1.0.101"
 
 [dev-dependencies]
-exonum-testkit = { version = "0.13.0-rc.2", path = "../../test-suite/testkit" }
-
 assert_matches = "1.3.0"
 bincode = "1.2.1"
 criterion = "0.3.0"
