+++ test-suite/rust-runtime-api/README.md
@@ -1,0 +1,12 @@
+# exonum-rust-runtime-api-tests
+
+[![Travis Build Status](https://img.shields.io/travis/exonum/exonum/master.svg?label=Linux%20Build)](https://travis-ci.com/exonum/exonum)
+[![License: Apache-2.0](https://img.shields.io/github/license/exonum/exonum.svg)](https://github.com/exonum/exonum/blob/master/LICENSE)
+![rust 1.36.0+ required](https://img.shields.io/badge/rust-1.36.0+-blue.svg?label=Required%20Rust)
+
+This is an internal crate for testing HTTP API endpoints of the `exonum-rust-runtime` crate.
+
+## License
+
+`exonum-rust-runtime-api-tests` is licensed under the Apache License (Version 2.0).
+See [LICENSE](LICENSE) for details.
