+++ test-suite/rust-runtime-api/src/proto/test_service.proto
