+++ test-suite/rust-runtime-api/src/service.rs
